backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
google-generativeai==0.8.3
PyJWT==2.8.0
cryptography==42.0.5
requests==2.32.5

